Move to less duplication in data-access ui
