+++ packages/ui-components/src/DataTable/utils.ts
@@ -1,5 +1,6 @@
 import { parseList, isString, uniq } from '@terascope/utils';
 import { ParsedSort, SortDirection, RowMapping } from './interfaces';
+import { ResolvedUser } from '../interfaces';
 
 export function parseSortBy(sort: string, defaultSort = 'created:asc'): ParsedSort {
     const parts = (sort || defaultSort).split(':');
@@ -17,10 +18,7 @@ export function uniqIntArray(arr: number[]) {
     return uniq(arr).sort((a, b) => a - b);
 }
 
-export function getSortDirection(
-    field: string,
-    sortBy: ParsedSort
-): 'ascending' | 'descending' {
+export function getSortDirection(field: string, sortBy: ParsedSort): 'ascending' | 'descending' {
     const none: any = null;
     if (sortBy.field !== field) return none;
     if (sortBy.direction === 'asc') return 'ascending';
@@ -29,12 +27,11 @@ export function getSortDirection(
 }
 
 export function formatRegexQuery(query: string, searchFields: string[]) {
-    const fields = searchFields.map((field) => `${field}.text`);
-    const fieldList = fields.map((val) => `${val}:/.*${query}.*/`);
+    const fields = searchFields.map(field => `${field}.text`);
+    const fieldList = fields.map(val => `${val}:/.*${query}.*/`);
     return fieldList.join(' OR ');
 }
 
-export function canSelectFn(rowMapping: RowMapping) {
-    return (record: any): boolean =>
-        rowMapping.canRemove ? rowMapping.canRemove(record) : true;
+export function canSelectFn(rowMapping: RowMapping, authUser?: ResolvedUser) {
+    return (record: any): boolean => (rowMapping.canRemove ? rowMapping.canRemove(record, authUser) : true);
 }
