+++ packages/ui-core/src/components/data-access/Users/List.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ModelList from '../ModelList';
+
+const List: React.FC = () => {
+    return <ModelList model="User" />;
+};
+
+export default List;
