+++ packages/ui-core/src/service-worker.ts
@@ -45,7 +45,9 @@ export function register(config?: Config) {
                 // service worker/PWA documentation.
                 navigator.serviceWorker.ready.then(() => {
                     // tslint:disable-next-line: no-console
-                    console.log('This web app is being served cache-first by a service ' + 'worker. To learn more, visit https://bit.ly/CRA-PWA');
+                    console.log(
+                        'This web app is being served cache-first by a service worker. To learn more, visit https://bit.ly/CRA-PWA'
+                    );
                 });
             } else {
                 // Is not localhost. Just register service worker
@@ -71,7 +73,10 @@ function registerValidSW(swUrl: string, config?: Config) {
                             // but the previous service worker will still serve the older
                             // content until all client tabs are closed.
                             // tslint:disable-next-line: no-console
-                            console.log('New content is available and will be used when all ' + 'tabs for this page are closed. See https://bit.ly/CRA-PWA.');
+                            console.log(
+                                'New content is available and will be used when all ' +
+                                    'tabs for this page are closed. See https://bit.ly/CRA-PWA.'
+                            );
 
                             // Execute callback
                             if (config && config.onUpdate) {
